+++ plugins/pastefromword/plugin.js
@@ -16,7 +16,8 @@
 		init: function( editor ) {
 			// Flag indicate this command is actually been asked instead of a generic pasting.
 			var forceFromWord = 0,
-				path = this.path;
+				path = this.path,
+				configInlineImages = editor.config.pasteFromWord_inlineImages === undefined ? true : editor.config.pasteFromWord_inlineImages;
 
 			editor.addCommand( 'pastefromword', {
 				// Snapshots are done manually by editable.insertXXX methods.
@@ -62,10 +63,13 @@
 				var data = evt.data,
 					dataTransferHtml = CKEDITOR.plugins.clipboard.isCustomDataTypesSupported ?
 						data.dataTransfer.getData( 'text/html', true ) : null,
+					// Required in paste from word image plugin (#662).
+					dataTransferRtf = CKEDITOR.plugins.clipboard.isCustomDataTypesSupported ?
+						data.dataTransfer.getData( 'text/rtf' ) : null,
 					// Some commands fire paste event without setting dataTransfer property. In such case
 					// dataValue should be used.
 					mswordHtml = dataTransferHtml || data.dataValue,
-					pfwEvtData = { dataValue: mswordHtml },
+					pfwEvtData = { dataValue: mswordHtml, dataTransfer: { 'text/rtf': dataTransferRtf } },
 					officeMetaRegexp = /<meta\s*name=(?:\"|\')?generator(?:\"|\')?\s*content=(?:\"|\')?microsoft/gi,
 					wordRegexp = /(class=\"?Mso|style=(?:\"|\')[^\"]*?\bmso\-|w:WordDocument|<o:\w+>|<\/font>)/,
 					isOfficeContent = officeMetaRegexp.test( mswordHtml ) || wordRegexp.test( mswordHtml );
@@ -106,6 +110,54 @@
 				// this event.
 				isLazyLoad && evt.cancel();
 			}, null, null, 3 );
+
+			// Paste From Word Image:
+			// RTF clipboard is required for embedding images.
+			// If img tags are not allowed there is no point to process images.
+			if ( CKEDITOR.plugins.clipboard.isCustomDataTypesSupported && configInlineImages ) {
+				editor.on( 'afterPasteFromWord', imagePastingListener );
+			}
+
+			function imagePastingListener( evt ) {
+				var pfw = CKEDITOR.plugins.pastefromword && CKEDITOR.plugins.pastefromword.images,
+					imgTags,
+					hexImages,
+					newSrcValues = [],
+					i;
+
+				// If pfw images namespace is unavailable or img tags are not allowed we simply skip adding images.
+				if ( !pfw || !evt.editor.filter.check( 'img[src]' ) ) {
+					return;
+				}
+
+				function createSrcWithBase64( img ) {
+					return img.type ? 'data:' + img.type + ';base64,' + CKEDITOR.tools.convertBytesToBase64( CKEDITOR.tools.convertHexStringToBytes( img.hex ) ) : null;
+				}
+
+				imgTags = pfw.extractTagsFromHtml( evt.data.dataValue );
+				if ( imgTags.length === 0 ) {
+					return;
+				}
+
+				hexImages = pfw.extractFromRtf( evt.data.dataTransfer[ 'text/rtf' ] );
+				if ( hexImages.length === 0 ) {
+					return;
+				}
+
+				CKEDITOR.tools.array.forEach( hexImages, function( img ) {
+					newSrcValues.push( createSrcWithBase64( img ) );
+				}, this );
+
+				// Assumption there is equal amount of Images in RTF and HTML source, so we can match them accordingly to existing order.
+				if ( imgTags.length === newSrcValues.length ) {
+					for ( i = 0; i < imgTags.length; i++ ) {
+						// Replace only `file` urls of images ( shapes get newSrcValue with null ).
+						if ( ( imgTags[ i ].indexOf( 'file://' ) === 0 ) && newSrcValues[ i ] ) {
+							evt.data.dataValue = evt.data.dataValue.replace( imgTags[ i ], newSrcValues[ i ] );
+						}
+					}
+				}
+			}
 		}
 
 	} );
@@ -154,6 +206,20 @@
  *
  * @since 3.1
  * @cfg {String} [pasteFromWordCleanupFile=<plugin path> + 'filter/default.js']
+ * @member CKEDITOR.config
+ */
+
+/**
+ * Flag decides whether embedding images pasted with Word content is enabled or not.
+ *
+ * **Note:** Please be aware that embedding images requires Clipboard API support, available only in modern browsers, that is indicated by
+ * {@link CKEDITOR.plugins.clipboard#isCustomDataTypesSupported} flag.
+ *
+ *		// Disable embedding images pasted from Word.
+ *		config.pasteFromWord_inlineImages = false;
+ *
+ * @since 4.8.0
+ * @cfg {Boolean} [pasteFromWord_inlineImages=true]
  * @member CKEDITOR.config
  */